backend/requirements.txt
Django==4.2.18
djangorestframework==3.15.2
django-allauth==65.3.1
dj-rest-auth[with_social]==7.0.1
psycopg2==2.9.10
drf-spectacular==0.28.0
django-cors-headers==4.6.0
paypalrestsdk==1.13.3
django-model-utils==5.0.0
django-hordak==1.17.0
django-money==3.5.3
pillow==11.1.0
django-phonenumber-field[phonenumbers]==8.0.0
faker
ruff
pre_commit==4.1.0
typing_extensions==4.12.2
django-filter==25.1
flake8
